move DocIDTransform class to proper file
