+++ transforms/universal/doc_id/dpk_doc_id/runtime.py
@@ -23,34 +23,12 @@ from data_processing.runtime.pure_python import (
 from data_processing.transform import TransformStatistics
 from data_processing.utils import ParamsUtils
 from dpk_doc_id.transform import (
-    DocIDTransformBase,
+    DocIDTransform,
     DocIDTransformConfigurationBase,
     IDGenerator,
     id_generator_key,
     start_id_key,
 )
-
-
-class DocIDTransform(DocIDTransformBase):
-    """
-    Implements schema modification of a pyarrow Table.
-    """
-
-    def __init__(self, config: dict[str, Any]):
-        """
-        Initialize based on the dictionary of configuration information.
-        """
-        # Make sure that the param name corresponds to the name used in apply_input_params method
-        super().__init__(config)
-        self.id_generator = config.get(id_generator_key, IDGenerator(config.get(start_id_key, 1)))
-
-    def _get_starting_id(self, n_rows: int) -> int:
-        """
-        Get starting ID
-        :param n_rows - number of rows in the table
-        :return: starting id for the table
-        """
-        return self.id_generator.get_ids(n_rows=n_rows)
 
 
 class DocIDTransformConfiguration(DocIDTransformConfigurationBase):
